Make solidity-match-variable-decls regexp's unhighlited matches shy.
This ensures that if the two optional matches in the regexp, `\\[ *[0-9]*\\]` or
`(regexp-opt solidity-variable-modifier 'words)` have no match, the name of the
variable being declared can be still be found via the absolute match index 2.
Prior to this change, a simple `address foo` would match the regexp, but the
expected match index for the variable, 4, would have no match, which could lead
to font-lock errors in some circumstances.

[Reference for shy matches](
https://www.gnu.org/software/emacs/manual/html_node/emacs/Regexp-Backslash.html#index-shy-group_002c-in-regexp-1000)